+++ test/App/AppTest.php
@@ -2,12 +2,6 @@
 namespace Gram\Test\App;
 
 use Gram\App\App;
-use Gram\Test\Middleware\DummyMw\TestMw1;
-use Gram\Test\Router\RouteMap;
-use Gram\Test\TestClasses\TestClass;
-use Nyholm\Psr7\Factory\Psr17Factory;
-use Nyholm\Psr7Server\ServerRequestCreator;
-use Pimple\Container;
 
 /**
  * Class AppTest
@@ -25,8 +19,6 @@ class AppTest extends AbstractAppTest
 	protected function setUp(): void
 	{
 		$this->initApp();
-
-		$this->app->init();
 	}
 
 }